+++ app/api/bins/[id]/route.js
@@ -1,9 +1,20 @@
 import dbConnect from "@/lib/dbConnect";
 import Bin from "@/models/bin";
-import collector from "@/models/collector";
+import Collector from "@/models/collector";
+import BinCollectionActivity from "@/models/binCollectionActivity";
+import { getServerSession } from "next-auth";
 
 export async function GET(req, { params }) {
   try {
+    const session = await getServerSession(req);
+
+    if (!session || !session.user) {
+      // Unauthorized if no session or user
+      return new Response(JSON.stringify({ message: "Unauthorized" }), {
+        status: 401,
+      });
+    }
+
     const { id } = await params;
     // Connect to the database
     await dbConnect();
@@ -27,8 +38,52 @@ export async function GET(req, { params }) {
 
 export async function DELETE(req, { params }) {
   try {
-    const { id } = await params;
+    const session = await getServerSession(req);
+
+    if (!session || !session.user) {
+      // Unauthorized if no session or user
+      return new Response(JSON.stringify({ message: "Unauthorized" }), {
+        status: 401,
+      });
+    }
+
     await dbConnect();
+
+    const collector = await Collector.findOne({ email: session.user.email });
+    if (!collector) {
+      return new Response(JSON.stringify({ error: "Collector not found" }), {
+        status: 404,
+        headers: { "Content-Type": "application/json" },
+      });
+    }
+
+    if (!collector.isAdmin) {
+      return new Response(JSON.stringify({ error: "Unauthorized" }), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      });
+    }
+
+    const { id } = await params;
+
+    //check if there are any routes, collections are associated with the bin
+    const bin = await Bin.findById(id);
+    if (!bin) {
+      return new Response(JSON.stringify({ error: "Bin not found" }), {
+        status: 404,
+        headers: { "Content-Type": "application/json" },
+      });
+    }
+
+    const collections = await BinCollectionActivity.find({ binId: id });
+    if (collections.length > 0 || bin.status === "processing") {
+      return new Response(
+        JSON.stringify({
+          error: "Bin cannot be deleted because it has collections",
+        }),
+        { status: 400, headers: { "Content-Type": "application/json" } }
+      );
+    }
 
     // Find and delete the collector by ID
     const deletedBin = await Bin.findByIdAndDelete(id);
@@ -62,7 +117,8 @@ export async function POST(req, { params }) {
     if (
       typeof isFull !== "boolean" ||
       latitude === undefined ||
-      longitude === undefined
+      longitude === undefined ||
+      (status && typeof status !== "string")
     ) {
       return new Response(JSON.stringify({ error: "Invalid data" }), {
         status: 400,
@@ -78,14 +134,26 @@ export async function POST(req, { params }) {
       });
     }
 
+    //check if the bin is already full or bin is in processing state, if so don't update the bin
+    if (bin.status === "processing") {
+      return new Response(
+        JSON.stringify({
+          success: false,
+          error:
+            "Bin is in processing state, unable update the bin untill the collection is completed",
+        }),
+        { status: 400 }
+      );
+    }
+
     // Update the bin data
     bin.isFull = isFull;
     bin.status = status || bin.status;
     if (latitude && longitude) {
       bin.location = { latitude, longitude };
     }
-    bin.updatedAt = new Date();
 
+    bin.updatedAt = new Date();
     await bin.save();
 
     if (global.io) {
@@ -95,14 +163,24 @@ export async function POST(req, { params }) {
 
     console.log("binUpdated");
     return new Response(
-      JSON.stringify({ message: "Bin updated successfully", bin }),
+      JSON.stringify({
+        success: true,
+        message: "Bin updated successfully",
+        bin,
+      }),
       { status: 200 }
     );
   } catch (error) {
     console.error("Error updating bin:", error);
-    return new Response(JSON.stringify({ error: "Failed to update bin" }), {
-      status: 500,
-      headers: { "Content-Type": "application/json" },
-    });
+    return new Response(
+      JSON.stringify({
+        success: false,
+        error: "Failed to update bin",
+      }),
+      {
+        status: 500,
+        headers: { "Content-Type": "application/json" },
+      }
+    );
   }
 }
